pyproject.toml
[tool.poetry]
name = "tdsql"
version = "0.1.0"
description = ""
authors = ["dr666m1 <skndr666m1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
